Refactoring to simplify reconstruction.
