+++ GIRF/GIRF_20200221_Duyn_method_coil2.mat
