+++ GIRF/__init__.py
@@ -1,0 +1,3 @@
+from .GIRF import apply_GIRF, calculate_matrix_pcs_to_dcs
+
+__all__ = ["apply_GIRF", "calculate_matrix_pcs_to_dcs"]
